Change Delta CheckpointEntryIterator ADD to use existing base column
The AddFileEntryExtractor was relying on a side effect of the Parquet
that merged columns with same name and different fields into a base
column. The proper way is to use a dereference projection, but this is
not needed here. Instead this code only needs one base column with the
correct field names.

With this change CheckpointFieldExtractor only need a single block.